Rework input stack in preparation for loading helix's config.toml
+ ground work for locating and reading helix config

This commit makes it so we generate bindings for every combination of modifiers and keys to allow the extension to decide at runtime if any possible binding should be active. Questionable approach? Perhaps. Effective approach? Absolutely. It also overhauls the internal representation of bindings/chords to be more streamlined to interact with.

Normal typing still goes through the type handler. Keys which work this way are the only ones which do not have bindings generated.
Escape and backspace are special cases because they do not get passed to the type handler ever, so they have bindings even without modifiers.
Bindings are toggled by generating a context var uniquely identifing each key combo e.g. binding the chord `ctrl+shift+f` results in `vscode.commands.executeCommand('setContext', "CSf", true);`
The same chord and identifier will be generated regardless of order of the modifiers, both simplifying everything and is needed for compatibility with helix config.

The generated bindings' commands emulate the input being routed to the active type handler. The main idea is to handle all the representations of input gathering and its edge cases in one system, and then use that to construct our own representation of input where we can pretend like everything goes through type handlers naturally. We can interpret every combination of key presses as being part of a chord. You can trivially do `["f"], ["ctrl", "f"]` as a chord, for example.
E.G. if someone's helix config says
```
[keys.normal]
"S-=" = "increment"
```
we can represent this, without any additonal plumbing not already found in this commit:
```
addBinding([actionFuncs["increment"]], [[Mode.Normal, ["shift", "="]]])
```

For a real result, the escape handler is no longer a specific binding, it is simply a one-key chord bound in every mode to a built in action, just like in helix:
```
	addBinding([actionFuncs.normal_mode], [
		[Mode.Insert, ["escape"]],
		[Mode.Normal, ["escape"]],
		[Mode.SearchInProgress, ["escape"]],
		[Mode.Select, ["escape"]],
		[Mode.Window, ["escape"]],
		[Mode.Visual, ["escape"]],
		[Mode.VisualLine, ["escape"]],
		[Mode.View, ["escape"]],
		[Mode.Match, ["escape"]]
	])```

Bindings are no longer represented as a list of functions which check a list of pressed keys as a pattern, instead bindings are stored in an array of modes, with nested bindings representing chords.
This allows us to use a single function which simply indexes into the binding list by our current mode, and continues indexing by the list of `HelixState.keysPressed`, producing 1 of 3 results each time:
* An object with a list of further keys (incomplete chord)
* A list of actions to execute (complete chord)
* Nothing (invalid/unconfigured chord)

TODO:
* Regex-based bindings are intended to be replaced with different type handlers and custom modes, such as how search was already implemented as its own mode (and now its own type handler as well).
* Reloading configs needs to reset all the binding identifiers in the extension context
* support for no_op functionality in helix's config
* Go through and remove/replace all the old extension commands
* Ensure we support all possible keys