Port regex actions to the new stack
Sorry for bad tabbing in previous commit.

The new Action type is now the only one (Action2 renamed to Action)

Add a "previous mode" functionality utilized by the match mode. This functionality is only guaranteed to be correct in execution paths which use `setPreviousMode` prior to using `enterPreviousMode`. Most paths don't do that.